Move purge modules to st2common so st2common doesn't need to depend on st2reactor.
Update affevted code and tests.
